requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
pyee==7.0.4
pyppeteer==0.2.2
python-dotenv==0.15.0
tqdm==4.53.0
urllib3==1.26.2
websockets==8.1
yarl==1.5.1
